use jsoncpp 1.8.3 instead of 0.x.y branch
Migrate to jsoncpp 1.x.x branch. This branch allows for use of exceptions and
other more up to date c++ features. Has teh ability to validate. Requires cmake >=3.1

+++ src/EPG_XML.h
@@ -47,7 +47,7 @@ class CEpg_Xml : public CEpgBase
   protected:
     Json::Value& FindJsonValue(Json::Value&, String, String);
     bool GzipInflate(const String&, String&);
-    int ParseDateTime(String&);
+    long int ParseDateTime(String&);
     bool XmlParse(HDHomeRunTuners::Tuner*, char*);
 
   private:
